pyproject.toml
[tool.poetry]
name = "text_watermarks"
version = "0.1.0"
description = "LN3 IT"
authors = ["Lex Chen"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^7.1.1"
pyqt5-tools = "^5.13.2"
pyinstaller = "^3.6"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
autopep8 = "^1.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
